deepen: Transaction Code Map
Replace two drifting copies of the SEC Form 4 transaction-code -> type map
with one deep Transaction Code Map module (lib/email/form4-transaction-codes.ts)
shared by the markdown extractor and the AI-JSON normalizer.

The extractor (form4-data-extractor.ts) and normalizer (form4-field-normalizer.ts)
are two render paths for the same Form 4 email; each carried its own copy of the
code->type map and they had drifted -- the extractor mapped 'A' to 'Award' while
the normalizer mapped it to 'Award/Grant'. A maintainer fixing the mapping in one
path would silently leave the other wrong.

- Add lib/email/form4-transaction-codes.ts as the single source of truth
  (canonical 20-code map; 'A' unified to 'Award/Grant', matching the primary
  AI-JSON path).
- form4-data-extractor and form4-field-normalizer both import it; delete their
  local copies. The normalizer's 12 codes already matched the canonical values
  exactly, so its behaviour is unchanged for those codes and additive for the
  rarer codes it previously lacked.
- The only behavioural change is the extractor's 'A' display label
  (Award -> Award/Grant), making the two paths consistent.

A third, semantically-divergent copy in lib/ai/parsers/response-parser.ts
(F->Disposition, C->Exercise, J/K->Transfer, W->Gift) is in a different pipeline
layer; reconciling it changes shipped Summary classification, so it is filed as
an architecture-gap rather than folded into this consolidation.

Tests: 316 form4-related tests pass (extractor, normalizer, classification,
derivative handling, response-parser, summarize-cached handlers). One assertion
updated to the unified 'Award/Grant' label. Added the Transaction Code Map term
to CONTEXT.md.
