+++ CONTEXT.md
@@ -44,3 +44,6 @@ An HMAC-signed token embedded in an outbound email link, covering both the feedb
 
 ### Historical Context
 A prompt enrichment that folds the most recent prior [Summary] entries for a ticker into the current [Filing] prompt so the model can reference continuity (e.g. "Q3 follows the Q2 guidance cut"). Capped at the last 3 summaries, each truncated to 1500 characters to bound token cost. Implemented as private functions inside the Summarize module (`lib/ai/summarize.ts`) — not a separate module — because the only caller is the summarization pipeline itself and the prior extraction-for-testability split offered no leverage. Failures are non-fatal: a Postgres error logs and proceeds without history.
+
+### Transaction Code Map
+The canonical mapping from a SEC Form 4 transaction code (single letters: S, P, A, F, M, J, K, …) to a human-readable type used in the insider-trading [Summary] and email. The atomic vocabulary for classifying what an insider did. Implemented as one deep module (`lib/email/form4-transaction-codes.ts`) behind a single `TRANSACTION_CODE_MAP` constant, shared by the markdown extractor (`form4-data-extractor.ts`) and the AI-JSON normalizer (`form4-field-normalizer.ts`) so the two Form 4 render paths can't drift. Replaces two copies that disagreed on the `A` label (`Award` vs `Award/Grant`; now uniformly `Award/Grant`). A third, semantically-divergent copy still lives in `lib/ai/parsers/response-parser.ts` (it maps `F`→Disposition, `C`→Exercise, `J`/`K`→Transfer, `W`→Gift, conflicting with SEC semantics); reconciling it changes shipped [Summary] classification, so it is tracked as an architecture-gap rather than folded in here.
